CoreML and TensorFlow object detection models seem to be working.
The accuracy of TensorFlow is not great, but at least it's running and given results.

This doesn't include the actual ML models yet, because I'm trying to figure out which ones to keep.

Both engines are abstracted behind a protocol that returns a Publisher of results, not sure I really like that approach, I may look at using something like PromiseKit.