Add GraphQL support (#23)
* Initial commit of GraphQL support

Add Apollo graphql library and some minimal functionality.

* squash

* SR Kit has basic integration with main application

Created the SRBackend protocol to support different actions on the GraphQL server.

* Added SRTestResult entity

Migrated the result view to at least compile with the new SR classes.

* squash

* Patient list and patient add are now working

Had to do some cleanups regarding the combine subscriptions, but seems to be working fine.

* Migrated order flow to GraphQL.

We're now storing orders, which are attached to a given SRPerson.

* Cleaned up handling of patient requests and results

Added the integration between the SRBackend and the existing TestResult flow.

The solution is terrible, we create an SRMonitor class which watches for changes to the ResultEntity table and forwards those along to the SRBackend. Ugh.

The benefit is that we don't

* Results are now being passed between LISKit and the application.

We have a totally hacky, incredibly terrible method of doing so, but it does seem to be working.

At least we're not blocking the main thread, I think....

* Results are flowing again.

For some reason, the SRTestResult struct is an absolute disaster.
No idea why it's that way, but my goodness, does anything work correctly?

Right now, we have results, but no dates.

* Disable tests, because who needs tests?

* Fix failing LISManager tests

Add the missing no-op backend.